Move shot allocation tests
And one minor bug fix
