+++ tests/test_expectation_samples.py
@@ -55,42 +55,6 @@ def test_measurement_basis_rotations_error():
 
 
 @pytest.mark.parametrize(
-    "method,max_shots_per_term,expected",
-    [
-        ("u", None, [66, 67, 67]),  # Control test; i.e. working normally
-        (None, None, [23, 168, 9]),  # Default arguments test
-        (None, 100, [75, 100, 25]),  # max_shots_per_term error
-        (None, 25, [75, 100, 25]),  # If max_shots_per_term is too small
-        (None, 1000, [23, 168, 9]),  # If max_shots_per_term is too large
-    ],
-)
-def test_allocate_shots(method, max_shots_per_term, expected):
-    hamiltonian = SymbolicHamiltonian(94 * Z(0) + Z(1) + 5 * X(0))  # Note that SymPy sorts the terms as X0 -> Z0 -> Z1
-    grouped_terms = measurement_basis_rotations(hamiltonian)
-    n_shots = 200
-    test_allocation = allocate_shots(
-        grouped_terms, method=method, n_shots=n_shots, max_shots_per_term=max_shots_per_term
-    )
-    # Might have the occasional off by one error, hence set the max allowed difference to be 1
-    assert max(abs(_i - _j) for _i, _j in zip(test_allocation, expected)) <= 1
-
-
-def test_allocate_shots_coefficient_edge_case():
-    """Edge cases of allocate_shots"""
-    hamiltonian = SymbolicHamiltonian(Z(0) + X(0))
-    grouped_terms = measurement_basis_rotations(hamiltonian)
-    n_shots = 1
-    assert allocate_shots(grouped_terms, n_shots=n_shots) in ([0, 1], [1, 0])
-
-
-def test_allocate_shots_input_validity():
-    hamiltonian = SymbolicHamiltonian(94 * Z(0) + Z(1) + 5 * X(0))
-    grouped_terms = measurement_basis_rotations(hamiltonian)
-    with pytest.raises(NameError):
-        _ = allocate_shots(grouped_terms, n_shots=1, method="wrong")
-
-
-@pytest.mark.parametrize(
     "gates_to_add,shot_allocation,expected",
     [
         ([gates.H(0)], [10, 0], 1.0),  # State vector: 1/sqrt(2)(|0> + |1>), Measuring X
